# api/gateway/analysis/api.yml
swagger: '2.0'
info:
  version: '1.0.0'  # API version, not the SAR code version
  title: panther-analysis-api
  description: Panther Python policies/rules
  contact:
    name: Panther Labs
    url: https://runpanther.io/about
    email: support@runpanther.io
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json

# The name of the CloudFormation resource for the Lambda handler function
...
paths:
...
  /test:
    post:
      operationId: TestPolicy
      summary: Test a policy against a set of unit tests
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/TestPolicy'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/TestPolicyResult'
        400:
          description: Bad request
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal server error
definitions:
  Error:
    type: object
    properties:
      message:
        description: Error message
        type: string
    required:
      - message

  ##### Create/Update/Modify #####
...
  TypeSet:
    description: List of resource/log types to which this policy applies
    type: array
    maxItems: 500
    uniqueItems: true
    items:
      type: string
      minLength: 1
  UnitTest:
    type: object
    properties:
      expectedResult:
        $ref: '#/definitions/testExpectedResult'
      name:
        $ref: '#/definitions/testName'
      resource:
        $ref: '#/definitions/testResource'
      resourceType:
        $ref: '#/definitions/testResourceType'
    required:
      - expectedResult
      - name
      - resource
      - resourceType
  TestSuite:
    type: array
    items:
      $ref: '#/definitions/UnitTest'
...
  TestPolicy:
    type: object
    properties:
      body:
        $ref: '#/definitions/body'
      resourceTypes:
        $ref: '#/definitions/TypeSet'
      tests:
        $ref: '#/definitions/TestSuite'
    required:
      - body
      - resourceTypes
      - tests
  TestPolicyResult:
    type: object
    properties:
      testSummary:
        $ref: '#/definitions/testSummary'
      testsPassed:
        $ref: '#/definitions/testsPassed'
      testsFailed:
        $ref: '#/definitions/testsFailed'
      testsErrored:
        $ref: '#/definitions/testsErrored'
    required:
      - testSummary
      - testsPassed
      - testsFailed
      - testsErrored

  ##### Suppress #####
...
  body:
    description: Python policy source code
    type: string
    minLength: 10
    maxLength: 1000000  # ~1 MB
...
  testExpectedResult:
    description: The expected outcome when running a unit test
    type: boolean
  testName:
    description: The name of a unit test
    type: string
    minLength: 1
  testResourceType:
    description: The resource or log type of an individual unit test
    type: string
    minLength: 1
  testResource:
    description: The resource (or log event) the unit test is running against
    type: string
    minLength: 2
...
  testSummary:
    description: True if all tests passed
    type: boolean
  testsPassed:
    description: List of names of tests that passed
    type: array
    items:
      type: string
  testsFailed:
    description: List of names of tests that failed
    type: array
    items:
      type: string
  testsErrored:
    description: List of names of tests that raised an error, along with their error message
    type: array
    items:
      $ref: '#/definitions/testErrorResult'
  testErrorResult:
    description: A combination of a test name and the error the test raised
    type: object
    properties:
      name:
        type: string
      errorMessage:
        type: string
...